+++ src/Modules/Settings/Settings.js
@@ -4,5 +4,7 @@ import '@/globals'
 import { vueApply, vueRegister } from '@/vue'
 import ProfileSettingsPage from './components/ProfileSettingsPage.vue'
 
-vueRegister({ ProfileSettingsPage })
+vueRegister({
+  ProfileSettingsPage,
+})
 vueApply('#profile-settings-page')