+++ tests/Unit/SettingsTransactionsTest.php
@@ -15,11 +15,13 @@ use Foodsharing\Modules\Settings\SettingsGateway;
 use Foodsharing\Modules\Settings\SettingsTransactions;
 use Foodsharing\Modules\Unit\UnitGateway;
 use Foodsharing\Permissions\SettingsPermissions;
+use Foodsharing\RestApi\Models\Settings\EmailChangeRequest;
 use Foodsharing\Utility\EmailHelper;
 use PHPUnit\Framework\MockObject\Exception;
 use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Tests\Support\UnitTester;
+use ValueError;
 
 class SettingsTransactionsTest extends Unit
 {
@@ -104,5 +106,147 @@ class SettingsTransactionsTest extends Unit
 
         // Old from DB
         $this->assertEquals(null, $this->transaction->getOption(UserOptionType::ACTIVITY_LISTINGS));
+    }
+
+    public function testStartRequestEMailChange()
+    {
+        $foodsaver = $this->tester->createStoreCoordinator('NeedThePassword', ['option' => '']);
+
+        $this->tester->cantSeeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id']]);
+        $this->session->expects($this->any())->method('id')->willReturn($foodsaver['id']);
+
+        $changeRequest = new EmailChangeRequest();
+        $changeRequest->email = 'Next@example.de';
+        $changeRequest->password = 'NeedThePassword';
+
+        // Action change of E-Mail
+        $this->transaction->requestEmailChange($changeRequest);
+
+        // Check creation of change token
+        $this->tester->seeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id'], 'newmail' => $changeRequest->email]);
+
+        $this->tester->expectNumMails(2);
+        $mails = $this->tester->getMails();
+
+        // Check Notification on old Mail address
+        $oldAddressMail = current(array_filter($mails, static function ($item) use ($foodsaver) {
+            return $item->headers->to == $foodsaver['email'];
+        }));
+        $this->tester->assertNotFalse($oldAddressMail, 'Mail not found');
+        $this->tester->assertRegExp("/\/user\/current\/settings\/email\/verifyAbort\?token\=[0-9a-f]+/m", $oldAddressMail->html, 'Link is missing');
+
+        // Check Notification with verification token in new E-Mail address
+        $newAddressMail = current(array_filter($mails, static function ($item) use ($changeRequest) {
+            return $item->headers->to == strtolower($changeRequest->email);
+        }));
+        $this->tester->assertNotFalse($newAddressMail, 'Mail not found');
+        $this->tester->assertRegExp("/\/user\/current\/settings\/email\/verify\?token\=[0-9a-f]+/m", $newAddressMail->html, 'Link is missing');
+    }
+
+    public function testCancelRequestEMailChange()
+    {
+        $foodsaver = $this->tester->createStoreCoordinator('NeedThePassword', ['option' => '']);
+
+        $this->tester->cantSeeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id']]);
+        $this->session->expects($this->any())->method('id')->willReturn($foodsaver['id']);
+
+        $changeRequest = new EmailChangeRequest();
+        $changeRequest->email = 'Next1@example.de';
+        $changeRequest->password = 'NeedThePassword';
+
+        // Action change of E-Mail
+        $this->transaction->requestEmailChange($changeRequest);
+        $this->tester->expectNumMails(2);
+        $mails = $this->tester->getMails();
+
+        // Find token for cancel in old e-mail
+        $oldAddressMail = current(array_filter($mails, static function ($item) use ($foodsaver) {
+            return $item->headers->to == $foodsaver['email'];
+        }));
+        $this->tester->assertNotFalse($oldAddressMail, 'Mail not found');
+        preg_match_all("/\/user\/current\/settings\/email\/verifyAbort\?token\=([0-9a-f]+)/m", $oldAddressMail->html, $tokens, PREG_SET_ORDER, 0);
+
+        $this->assertNotNull($tokens);
+        $token = $tokens[0][1];
+
+        $this->transaction->abortEMailChange($token);
+        $this->tester->dontSeeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id']]);
+        $this->tester->seeInDatabase('fs_foodsaver_change_history', ['fs_id' => $foodsaver['id'], 'object_name' => 'emailAbort', 'old_value' => $foodsaver['email'], 'new_value' => strtolower($changeRequest->email)]);
+
+        // E-Mail??
+    }
+
+    public function testVerificationOfEMailChangeRequest()
+    {
+        $foodsaver = $this->tester->createStoreCoordinator('NeedThePassword', ['option' => '']);
+
+        $this->tester->cantSeeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id']]);
+        $this->session->expects($this->any())->method('id')->willReturn($foodsaver['id']);
+
+        $changeRequest = new EmailChangeRequest();
+        $changeRequest->email = 'Next2@example.de';
+        $changeRequest->password = 'NeedThePassword';
+
+        // Action change of E-Mail
+        $this->transaction->requestEmailChange($changeRequest);
+        $this->tester->expectNumMails(2);
+        $mails = $this->tester->getMails();
+
+        // Find token for verify link in new e-mail
+        $newAddressMail = current(array_filter($mails, static function ($item) use ($changeRequest) {
+            return $item->headers->to == strtolower($changeRequest->email);
+        }));
+        $this->tester->assertNotFalse($newAddressMail, 'Mail not found');
+        preg_match_all("/\/user\/current\/settings\/email\/verify\?token\=([0-9a-f]+)/m", $newAddressMail->html, $tokens, PREG_SET_ORDER, 0);
+
+        $this->assertNotNull($tokens);
+        $token = $tokens[0][1];
+
+        $this->transaction->verifyAndCompleteEMailChange($token);
+        $this->tester->dontSeeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id']]);
+        $this->tester->seeInDatabase('fs_foodsaver', ['id' => $foodsaver['id'], 'email' => strtolower($changeRequest->email)]);
+        $this->tester->seeInDatabase('fs_foodsaver_change_history', ['fs_id' => $foodsaver['id'], 'object_name' => 'email', 'old_value' => $foodsaver['email'], 'new_value' => strtolower($changeRequest->email)]);
+
+        // E-Mail??
+    }
+
+    public function testVerificationOfEMailChangeRequestFailsEMailInUse()
+    {
+        $foodsaver = $this->tester->createStoreCoordinator('NeedThePassword', ['option' => '']);
+
+        $this->tester->cantSeeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id']]);
+        $this->session->expects($this->any())->method('id')->willReturn($foodsaver['id']);
+
+        $changeRequest = new EmailChangeRequest();
+        $changeRequest->email = 'Next2@example.de';
+        $changeRequest->password = 'NeedThePassword';
+
+        // Action change of E-Mail
+        $this->transaction->requestEmailChange($changeRequest);
+        $this->tester->expectNumMails(2);
+        $mails = $this->tester->getMails();
+
+        // Find token for verify link in new e-mail
+        $newAddressMail = current(array_filter($mails, static function ($item) use ($changeRequest) {
+            return $item->headers->to == strtolower($changeRequest->email);
+        }));
+        $this->tester->assertNotFalse($newAddressMail, 'Mail not found');
+        preg_match_all("/\/user\/current\/settings\/email\/verify\?token\=([0-9a-f]+)/m", $newAddressMail->html, $tokens, PREG_SET_ORDER, 0);
+
+        $this->assertNotNull($tokens);
+        $token = $tokens[0][1];
+
+        $foodsaver2 = $this->tester->createStoreCoordinator('NeedThePassword', ['option' => '', 'email' => strtolower($changeRequest->email)]);
+
+        try {
+            $this->transaction->verifyAndCompleteEMailChange($token);
+            $this->tester->assertFalse(true, 'Unexpected code executed');
+        } catch (ValueError $e) {
+        }
+        $this->tester->dontSeeInDatabase('fs_mailchange', ['foodsaver_id' => $foodsaver['id']]);
+        $this->tester->dontSeeInDatabase('fs_foodsaver', ['id' => $foodsaver['id'], 'email' => strtolower($changeRequest->email)]);
+        $this->tester->dontSeeInDatabase('fs_foodsaver_change_history', ['fs_id' => $foodsaver['id'], 'object_name' => 'email', 'old_value' => $foodsaver['email'], 'new_value' => strtolower($changeRequest->email)]);
+
+        // E-Mail??
     }
 }